Add override support of sizes
The possibility to use the 'size' property of a field to override its
predefined size is added. For example, this allows to set a size of 6
bytes to extract a UINT64. Also, it will allow to set sizes for types
without any predefined size values like RELATIVE_TIME.

Update the README accordingly.